Api change.

+++ common/src/main/java/com/hehehey/ghost/message/frontend/TaskProgress.java
@@ -14,12 +14,12 @@ public class TaskProgress {
     /**
      * 任务剩余未完成的网址
      */
-    private int remainingUrlCount;
+    private long remainingUrlCount;
 
     /**
      * 此次任务的当前数据量
      */
-    private int dataCount;
+    private long dataCount;
 
     public String getId() {
         return id;
@@ -29,19 +29,19 @@ public class TaskProgress {
         this.id = id;
     }
 
-    public int getRemainingUrlCount() {
+    public long getRemainingUrlCount() {
         return remainingUrlCount;
     }
 
-    public void setRemainingUrlCount(int remainingUrlCount) {
+    public void setRemainingUrlCount(long remainingUrlCount) {
         this.remainingUrlCount = remainingUrlCount;
     }
 
-    public int getDataCount() {
+    public long getDataCount() {
         return dataCount;
     }
 
-    public void setDataCount(int dataCount) {
+    public void setDataCount(long dataCount) {
         this.dataCount = dataCount;
     }
 }
